Made the newsletter design using the mobile first approach
Using this approach seems easier than using the desktop first approach.
Anyways, I had to learn about the hamburger issue of resizing the screen
and how the links underneath does not fit in. I bypassed this using the
event listener of resize to the window to set the width of the nav based
on the window size.