Design the SunnySide Website
Modified the css and js file to include all the needed details to make
the site responsive as needed. Anyways learnt something about visibility
and responsive image.
The the visibilitity just like display none, hides an element from the UI,
however, it still in the document flow making it still interactive unlike
diplay none.
Next, understood the responsive feature of img tag by utilizing the srcset
and sizes attribute, where the srcset contains the list of images to use with
a width descriptor (w) and the sizes defines media queries and the space to
occupy when the media propety is satisfied.

Note also that I was made to understand that when designing a hamburger menu
using the mobile first approach, the reason why the nav sometimes fail to fit
to the screen is because of an undefined width or max-width.