Add the Interactive feature for the pricing card
While the feature I stumbled making a drag and move effect for an
element, just basically trying to simulate a slider. I later came to
use the slider element for accessibilty reasons where I later later
learnt how to style it differently using the webkit apperance and moz
appearance. However, I was able to restrict the range element using js.
Now, for the custom slider, I used mouse down and mouse up feature, although
the mouseup feature was set on the document. Lastly, to add movement, I had
to utilize a BrowserAPI called getBoundingClientRect which is simple words
give the dimension and position of an element it is called upon.