Design the loopStudio landing page
This build was quite great, made some progress in understanding how a
picture element works and was able to utilize the backdrop filter of
tailwind. Lastly had a little confusion on my navbar as per transition
so I left it plain. Will make sure to work on this in some other project.