Made the Social Media Dashboard
Learnt that gradient cannot be attached to border so have to have
implement a pseudo border using the before pseudo element.

Next learnt how to implement a dark mode on tailwind by utilizing
the tailwind configuration of darkMode and the css pseudo dark: element
created by tailwind. Note, however, I gave the end user the luxury of
being able to change his or her own theme by using the a function that
does this following: first checks if we have the theme attribute set to
dark in local storage or theme is not in local storage and prefer color
of OS is dark the we can change to dark mode otherwise light. Note that
the change is by adding or removing a class with value of dark to html
element which is repesented by documentElement of the DOM, and all this
is tied to the dark: to make effect on element. Note, the darkMode: selector
settings is to allow a user to change his theme by the click of a button.